refactor: break out database config tests to separate file.
