+++ packages/serverpod_shared/test/config_test.dart
@@ -129,10 +129,6 @@ apiServer:
       expect(config.serviceSecret, 'LONG_PASSWORD_THAT_IS_REQUIRED');
     });
 
-    test('Database configuration is null.', () {
-      expect(config.database, isNull);
-    });
-
     test('Redis configuration is null.', () {
       expect(config.redis, isNull);
     });
@@ -216,100 +212,6 @@ maxRequestSize: 1048576
     );
 
     expect(config.maxRequestSize, 1048576);
-  });
-
-  test(
-      'Given a Serverpod config with database configuration without password when loading from Map then exception is thrown.',
-      () {
-    var serverpodConfig = '''
-apiServer:
-  port: 8080
-  publicHost: localhost
-  publicPort: 8080
-  publicScheme: http
-database:
-  host: localhost
-  port: 5432
-  name: testDb
-  user: test
-''';
-
-    expect(
-      () => ServerpodConfig.loadFromMap(
-        runMode,
-        serverId,
-        passwords,
-        loadYaml(serverpodConfig),
-      ),
-      throwsA(isA<Exception>().having(
-        (e) => e.toString(),
-        'message',
-        equals('Exception: Missing database password.'),
-      )),
-    );
-  });
-
-  test(
-      'Given a Serverpod config with database configuration missing required field when loading from Map then exception is thrown.',
-      () {
-    var serverpodConfig = '''
-apiServer:
-  port: 8080
-  publicHost: localhost
-  publicPort: 8080
-  publicScheme: http
-database:
-  host: localhost
-  name: testDb
-  user: test
-''';
-
-    expect(
-      () => ServerpodConfig.loadFromMap(
-        runMode,
-        serverId,
-        {...passwords, 'database': 'password'},
-        loadYaml(serverpodConfig),
-      ),
-      throwsA(isA<Exception>().having(
-        (e) => e.toString(),
-        'message',
-        equals(
-            'Exception: database is missing required configuration for port.'),
-      )),
-    );
-  });
-
-  test(
-      'Given a Serverpod config with database configuration when loading from Map then database configuration is set.',
-      () {
-    var serverpodConfig = '''
-apiServer:
-  port: 8080
-  publicHost: localhost
-  publicPort: 8080
-  publicScheme: http
-database:
-  host: localhost
-  port: 5432
-  name: testDb
-  user: test
-''';
-
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      loadYaml(serverpodConfig),
-    );
-
-    expect(config.database?.host, 'localhost');
-    expect(config.database?.port, 5432);
-    expect(config.database?.name, 'testDb');
-    expect(config.database?.user, 'test');
-    expect(config.database?.password, 'password');
-    expect(config.database?.requireSsl, isFalse);
-    expect(config.database?.isUnixSocket, isFalse);
   });
 
   test(
@@ -682,128 +584,6 @@ redis:
   });
 
   test(
-      'Given a Serverpod config with only the api server configuration but the environment variables containing the config for the database when loading from Map then the database config is created.',
-      () {
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      {
-        'apiServer': {
-          'port': 8080,
-          'publicHost': 'localhost',
-          'publicPort': 8080,
-          'publicScheme': 'http',
-        },
-      },
-      environment: {
-        'SERVERPOD_DATABASE_HOST': 'localhost',
-        'SERVERPOD_DATABASE_PORT': '5432',
-        'SERVERPOD_DATABASE_NAME': 'serverpod',
-        'SERVERPOD_DATABASE_USER': 'admin',
-      },
-    );
-
-    expect(config.database?.host, 'localhost');
-    expect(config.database?.port, 5432);
-    expect(config.database?.name, 'serverpod');
-    expect(config.database?.user, 'admin');
-  });
-
-  test(
-      'Given a Serverpod config map with half the values and the environment variables the other half for the database when loading from Map then configuration then the database config is created',
-      () {
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      {
-        'apiServer': {
-          'port': 8080,
-          'publicHost': 'localhost',
-          'publicPort': 8080,
-          'publicScheme': 'http',
-        },
-        'database': {
-          'port': 5432,
-          'name': 'serverpod',
-        },
-      },
-      environment: {
-        'SERVERPOD_DATABASE_HOST': 'localhost',
-        'SERVERPOD_DATABASE_USER': 'admin',
-      },
-    );
-
-    expect(config.database?.host, 'localhost');
-    expect(config.database?.port, 5432);
-    expect(config.database?.name, 'serverpod');
-    expect(config.database?.user, 'admin');
-  });
-
-  test(
-      'Given a Serverpod config map with all the values and the environment variables for the database when loading from Map then the config is overridden by the environment variables.',
-      () {
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      {
-        'apiServer': {
-          'port': 8080,
-          'publicHost': 'localhost',
-          'publicPort': 8080,
-          'publicScheme': 'http',
-        },
-        'database': {
-          'host': 'localhost',
-          'port': 5432,
-          'name': 'serverpod',
-          'user': 'admin',
-        },
-      },
-      environment: {
-        'SERVERPOD_DATABASE_HOST': 'remotehost',
-        'SERVERPOD_DATABASE_PORT': '5433',
-        'SERVERPOD_DATABASE_NAME': 'remote_serverpod',
-        'SERVERPOD_DATABASE_USER': 'remote_admin',
-      },
-    );
-
-    expect(config.database?.host, 'remotehost');
-    expect(config.database?.port, 5433);
-    expect(config.database?.name, 'remote_serverpod');
-    expect(config.database?.user, 'remote_admin');
-  });
-
-  test(
-      'Given a Serverpod config with only the api server configuration but the environment variables containing the optional database variable require ssl then the database config takes the value from the env.',
-      () {
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      {
-        'apiServer': {
-          'port': 8080,
-          'publicHost': 'localhost',
-          'publicPort': 8080,
-          'publicScheme': 'http',
-        },
-      },
-      environment: {
-        'SERVERPOD_DATABASE_HOST': 'localhost',
-        'SERVERPOD_DATABASE_PORT': '5432',
-        'SERVERPOD_DATABASE_NAME': 'serverpod',
-        'SERVERPOD_DATABASE_USER': 'admin',
-        'SERVERPOD_DATABASE_REQUIRE_SSL': 'true',
-      },
-    );
-
-    expect(config.database?.requireSsl, true);
-  });
-
-  test(
       'Given a Serverpod config with only the api server configuration but the environment variables containing the optional database variable require ssl set to an invalid value then an exception is thrown.',
       () {
     expect(
@@ -834,33 +614,6 @@ redis:
             'Exception: Invalid value (INVALID) for SERVERPOD_DATABASE_REQUIRE_SSL.'),
       )),
     );
-  });
-
-  test(
-      'Given a Serverpod config with only the api server configuration but the environment variables containing the optional database variable isUnixSocket then the database config takes the value from the env.',
-      () {
-    var config = ServerpodConfig.loadFromMap(
-      runMode,
-      serverId,
-      {...passwords, 'database': 'password'},
-      {
-        'apiServer': {
-          'port': 8080,
-          'publicHost': 'localhost',
-          'publicPort': 8080,
-          'publicScheme': 'http',
-        },
-      },
-      environment: {
-        'SERVERPOD_DATABASE_HOST': 'localhost',
-        'SERVERPOD_DATABASE_PORT': '5432',
-        'SERVERPOD_DATABASE_NAME': 'serverpod',
-        'SERVERPOD_DATABASE_USER': 'admin',
-        'SERVERPOD_DATABASE_IS_UNIX_SOCKET': 'true',
-      },
-    );
-
-    expect(config.database?.isUnixSocket, true);
   });
 
   test(
